requirements.txt
langchain-community
langchain-openai
faiss-cpu
python-dotenv
numpy
